requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-chroma>=0.0.2
openai>=1.0.0
chromadb>=0.4.14
pypdf>=3.15.1
python-dotenv>=1.0.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2
pydantic>=2.4.2
rich>=13.5.2
typer>=0.9.0 